Update equilibriumPoint Q no 2.txt
Question no 2:
Given an array A of n positive numbers. The task is to find the first Equilibrium Point in the array. Equilibrium Point in an array is a position such that the sum of elements before it is equal to the sum of elements after it.
The task is to complete the function equilibriumPoint() which takes the array and n as input parameters and returns the point of equilibrium. Return -1 if no such point exists.
Expected Time Complexity: O(n) Expected Auxiliary Space: O(1)
Constraints:1 <= n <= 1061 <= A[i] <= 108.